requirements.txt
Django==5.1.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-filter==24.3
django-cors-headers==4.6.0
django-extensions==3.2.3
django-debug-toolbar==4.4.4
psycopg2-binary==2.9.9
Pillow==12.0.0
inflect
jinja2

